Add information displayed in misfit plots
These changes are regarding the plugin Observation Analyzer.
We display the realization that specific data points come from when
hovering in the univariate misfit boxplots, and also display the
ensemble for all hovertexts and in the legend, grouping the legend
entries for different ensembles by date. We also display the realization
in hover text for the misfit summary histogram plots.