Refactor select_first function
Instead of clicking a coordinate, just pick the list of options
and click the first one. The motivation for the refactoring is
tests started breaking due to a change in selenium
move_to_element_with_offset() which currently tries to use the
top left corner of the element as the origin; in Selenium 4.3
it will use the in-view center point of the element as the origin.