Make sorting optional in response correlation viewer
When comparing how parameters affect responses, when switching between
different responses, the rows of the heatmap and tornado plots,
representing different parameters, were shuffled: parameters were sorted
based on level of correlation.

This behaviour can be confusing when trying to keep track of specific
parameters, hence we make it possible to turn it off.

Cf. #304