Fix response overview plot to survive reloads
In the response correlation viewer, the response overview plot was not
getting redrawn on browser refresh. Using a data store as input for the
responsible callback function fixes the problem.

When trying to add a regression test, we were faced with either
introducing flakiness via the occasional "stale element" error from
selenium, or a moot test that didn't catch the buggy behaviour when
removing the flakiness with a wait. So we restrict ourselves to manual
testing for this particular change.