Improve display of active variables
In the response correlation plugin, one may choose an active response
and parameter among the selected responses and paramteres, and an active
index. The way the information which variable is currently active is
displayed is made more prominent and compact in this change.

To this purpose, a bootstrap css file is included among the assets to be
able to leverage the components included therein.

Simultaneously, we move the options that affect the heatmap and tornado
plots together to the bottom of the page, and rearrange them a bit.