Fix visuals in response comparison view
Sometimes in the response comparison view, I believe perhaps due to a
glitch / race condition in react / dash, when one removes one of two
response plots while parameter plots are displayed, the response plots
are resized in an unfortunate way, and cover up parts of the parameter
plot titles. We fix the height, which should rectify that issue.

We also make labels completely visible in the parameter comparison view.

Closes #378