Make license consistent in PyPI and README
The wrong license was hardcoded into the README badge. This update
instead gets it from PyPI, which in turn uses the project metadata.
Since the metadata was not set, I added it to setup.py -- so it won't
appear in PyPI until the next release. After that, all will be good.